Issue #141,154 Issue #141,154 Ticket #154 is the 11 Dalton Shift.  A large number of these commits are put together for these two features.  To the clsMassTagDB we added the ability to provide a dalton shift when retrieving the items.
Ticket #141
To the cluster Grid we added a SMART Scores column.
To the wizard pages we added SMART parameter windows and buttons.  The ability to select what type of peak matching to perform (Peptide ID) SMART or Standard Peak Matching.

Both:
To the summary pages we added fields that have SMART summary tables, along with FDR calculations in the peak matching results tab of the Results Summary.